requirements.txt
scipy
boto3==1.12.19
botocore==1.15.19
certifi==2019.11.28
chardet==3.0.4
docutils==0.15.2
fsspec==0.6.2
idna==2.9
Jinja2==2.11.1
jmespath==0.9.5
MarkupSafe==1.1.1
numpy==1.18.1
pandas==1.0.1
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
requests-cache==0.5.2
s3fs==0.4.0
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
